fix(report): rename system-image-legend file to fix typo
Corrects a typo in the filename (compoent -> component) and updates
the two imports that referenced it.
