+++ apps/frontend/src/view/report/components/system-image-legend.report.component.tsx
